requirements.txt
torch
ultralytics
opencv-python
numpy
deep-sort-realtime
python-dotenv
google-generativeai
tqdm